Compatible with Laravel 6 (#5)
* Compatible with Laravel 6

* move method to abstract class.

* Update Readme.md
