+++ composer.json
@@ -9,15 +9,21 @@
     "homepage" : "https://github.com/kra8/laravel-snowflake",
     "license": "MIT",
     "require": {
-        "php": ">=7.0.0"
+        "php": "^7.0|^7.2",
+        "illuminate/support": "^6.0"
     },
     "require-dev" : {
-		"phpunit/phpunit": "~6.0",
-        "orchestra/testbench": "~3.0"
+		"phpunit/phpunit": "~8.0",
+        "orchestra/testbench": "~4.0"
     },
     "autoload": {
         "psr-4": {
             "Kra8\\Snowflake\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Kra8\\Snowflake\\Test\\": "tests/"
         }
     },
     "extra": {
